UBUNTU: Enable both IEEE1394 (Firewire) stacks as modules
Bug: #276463

The kernel now has two competing firewire (IEEE1394) stacks.  We are
following upstream recomendations and enabling the older legacy stack
for general use.  There are a number of requests to include the new
stack as well (as blacklisted modules) to allow simple switching to the
new stack for those who like to test the bleeding edge.

Enable the new firewire stack.  This does need a corresponding
module-init-tools change to ensure we do not load these.

This effectively reverts 'UBUNTU: Build in IEEE1394 OHCI1394'

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>