UBUNTU: update kernel bootloader recommends: to prefer grub
Bug: #314004

The kernel packages recommend either lilo or grub.  The order of these
indicates a preference.  Currently we recommend lilo over grub which is
contrary to the default in debian-installer.  From Intrepid recommendations
are automatically installed which causes the default desktop installs to
install lilo.  When the installer selects grub it then installs that as
well leading us to have both installled (a bug in the installer as it
should remove the redundant one).  Even with this installer bug fixed
we will always install the first one listed on the kernel Recommends:
and then swap it to the other should that be selected, it is logical to
default to the one which is most commonly used; grub.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>