UBUNTU: SAUCE: Mask off garbage in Dell WMI scan code data
From Mario Limonciello: I'm attaching a patch that has been accepted into -mm, but I want to
make sure it's available for Karmic in case it doesn't make it in for
2.6.30 (or whatever kernel karmic ends up with).

---------------
In debugging with some future machines that actually contain BIOS level
support for dell-wmi, I've determined that the upper half of the data that
comes back from wmi_get_event_data may sometimes contain extra information
that isn't currently relevant when pulling scan codes out of the data.
This causes dell-wmi to improperly respond to these events.

Cc: Matthew Garrett <mjg@redhat.com>
Cc: Dmitry Torokhov <dtor@mail.ru>
Cc: Len Brown <len.brown@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>