UBUNTU: SAUCE: Add MODULE_ALIAS for Dell WMI module
From Mario Limonciello: Until the sysfs WMI interface is ready, there is no way
to automatically determine when to load dell-wmi.  The equivalent module for Acer
laptops, acer-wmi has a workaround in place to match on the DMI tables for the machine.
This type of solution works on for Dell as well and gets the module loading when necessary.

Upstream is not in favor of applying this patch, but rather waiting for the sysfs
interface to be ready. Until it is, this provides a solution to ensure machines that
support WMI hotkeys are functional.

OriginalAuthor: mario_limonciello@dell.com

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>