UBUNTU: [Config] ports - Do not update ports kernel configurations by default
Since the kernel team are not responsible for ports kernels, they don't want
to deal with updating ports kernel configs. When debian/rules updateconfigs or
debian/rules editconfigs are run, only the supported architectures, amd64,
armel, i386 and lpia are updated. When debian/rules updateconfigs or
debian/rules editconfigs are run, all architectures are updated.

This is the best way I can currently think of to ensure the kernel team doesn't
have to deal with architectures they don't want to. The supported architecture
list is currently hard coded in the kernelconfig script, however if needed,
this can easily be set elsewhere.

Signed-off-by: Luke Yelavich <themuso@ubuntu.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>