UBUNTU: [Config] disable CONFIG_DM_RAID45 on powerpc
The likelyhood of users wanting to access dmraid arrays on powerpc is low.
The likelyhood of users wanting to access a dmraid RAID5 array on powerpc
is extremely rare, so I think its safe to turn off this module. This is
due to the fact that dmraid is a PC BIOS function, and not something that
powerpc hardware vendors would consider offering, at least so far as I
know.

Signed-off-by: Luke Yelavich <themuso@ubuntu.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>