UBUNTU: SAUCE: AppArmor: Fix profile attachment for regexp based profile names
BugLink: http://bugs.launchpad.net/bugs/419308

Part a & b
Fix profile attachment for profiles that specify an xmatch dfa.  Just testing
if the state > DFA_START is not sufficient to determine if the state is an
accepting state, as there are transition states that are not accept states.
Instead test for the MAY_EXEC permission which is set for valid matches.

Part c & d of lp #419308:
The attachment semantics for px, cx are currently wrong.  They do correct
attachment for profiles with names that are exact matches but the
attachment fails for profiles names that contain regular expressions.

eg. Given the following profiles

/bin/** { } and /bin/foo { }

an unconfined process correctly attaches to either profile but for a
confined process with x transition rules.
  /bin/foo px,   # correctly attaches
  /bin/bar px,   # does not attach to /bin/** as it should.

Use the system attachment function, instead of searching for the profile.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>