UBUNTU: [Upstream] ACPI: Add Thinkpad X300 & Thinkpad X301 to OSI(Linux) white-list
BugLink: http://bugs.launchpad.net/bugs/281732

Fix mute key for X300 series Thinkpads

acpi_osi=Linux helps the mute button work properly by sending Linux
a mute key press. This is the same behavior as seen with the Thinkpad
T61 already on the white list.

The other issue fixed is that with the X300 & X301 the mute button
only controls the speakers and not the headphone jack.

One side effect of this is that the mute button will no longer glow
orange when speakers are muted. If this new behavior is not desired
then user can place acpi_osi=!Linux on kernel command line.

Signed-off-by: Jerone Young <jerone.young@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>