UBUNTU: [Upstream] acerhdf: Limit modalias matching to supported boards
BugLink: http://bugs.launchpad.net/bugs/435958

Upstream status: acked, but not accepted, yet

The module alias currently matches any Acer computer but when loaded the
BIOS checks will only succeed on Aspire One models. This causes a invalid
BIOS warning for all other models (seen on Aspire 4810T).
This is not fatal but worries users that see this message. Limiting the
moule alias to models starting with AOA. Probably the same change should be
made for the two other entries.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Borislav Petkov <petkovbb@gmail.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>